# caresoft_core/caresoft_core/Connected Services/CoreWebApi/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "caresoft_core",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ReservaServicio/toggle-state/{id}": {
      "put": {
        "tags": [
          "ReservaServicio"
        ],
        "operationId": "toggle-state",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Usuario/get/{codigoODocumento}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "operationId": "get8",
        "parameters": [
          {
            "name": "codigoODocumento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Usuario/add": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "operationId": "add16",
        "parameters": [
          {
            "name": "UsuarioCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Documento",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "UsuarioContra",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "TipoDocumento",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NumLicenciaMedica",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Apellido",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Genero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "FechaNacimiento",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Telefono",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Correo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direccion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Rol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/update": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "operationId": "update16",
        "parameters": [
          {
            "name": "UsuarioCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Documento",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "UsuarioContra",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "TipoDocumento",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NumLicenciaMedica",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Apellido",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Genero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "FechaNacimiento",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Telefono",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Correo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direccion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Rol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Usuario/delete/{codigoOdocumento}": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "operationId": "delete16",
        "parameters": [
          {
            "name": "codigoOdocumento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Usuario/cuenta/{codigoOdocumento}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "operationId": "cuenta",
        "parameters": [
          {
            "name": "codigoOdocumento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CuentumDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuentumDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuentumDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Usuario/toggle-state-cuenta/{codigoOdocumento}": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "operationId": "toggle-state-cuenta",
        "parameters": [
          {
            "name": "codigoOdocumento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CuentumDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idCuenta": {
            "type": "integer",
            "format": "int32"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "estado": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UsuarioDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "usuarioCodigo": {
            "type": "string",
            "nullable": true
          },
          "documento": {
            "type": "string",
            "nullable": true
          },
          "usuarioContra": {
            "type": "string",
            "nullable": true
          },
          "tipoDocumento": {
            "type": "string",
            "nullable": true
          },
          "numLicenciaMedica": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "genero": {
            "type": "string",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "rol": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}